move a number of commands out of cmd.rs
