+++ src/command_processor.rs
@@ -158,7 +158,6 @@ impl CommandProcessor {
 
     fn get_command_vec() -> Vec<Box<dyn Cmd>> {
         let commands: Vec<Box<dyn Cmd>> = vec![
-            Box::new(crate::cmd::Quit::new()),
             Box::new(crate::cmd::Context::new()),
             Box::new(crate::cmd::Contexts::new()),
             Box::new(crate::cmd::Range::new()),
@@ -167,22 +166,23 @@ impl CommandProcessor {
             Box::new(crate::cmd::ReplicaSets::new()),
             Box::new(crate::cmd::StatefulSets::new()),
             Box::new(crate::cmd::ConfigMaps::new()),
-            Box::new(crate::cmd::Namespace::new()),
             Box::new(crate::cmd::Logs::new()),
             Box::new(crate::cmd::Describe::new()),
             Box::new(crate::cmd::Exec::new()),
             Box::new(crate::cmd::Events::new()),
-            Box::new(crate::cmd::Clear::new()),
-            Box::new(crate::cmd::EnvCmd::new()),
-            Box::new(crate::cmd::SetCmd::new()),
             Box::new(crate::cmd::Delete::new()),
-            Box::new(crate::cmd::UtcCmd::new()),
             Box::new(crate::cmd::Secrets::new()),
             Box::new(crate::cmd::PortForward::new()),
             Box::new(crate::cmd::PortForwards::new()),
             Box::new(crate::cmd::Jobs::new()),
-            Box::new(crate::cmd::Alias::new()),
-            Box::new(crate::cmd::Unalias::new()),
+            Box::new(crate::command::alias::Alias::new()),
+            Box::new(crate::command::alias::Unalias::new()),
+            Box::new(crate::command::click::Clear::new()),
+            Box::new(crate::command::click::EnvCmd::new()),
+            Box::new(crate::command::click::Quit::new()),
+            Box::new(crate::command::click::SetCmd::new()),
+            Box::new(crate::command::click::UtcCmd::new()),
+            Box::new(crate::command::namespaces::Namespace::new()),
             Box::new(crate::command::namespaces::Namespaces::new()),
             Box::new(crate::command::nodes::Nodes::new()),
             Box::new(crate::command::volumes::PersistentVolumes::new()),
