+++ src/completer.rs
@@ -311,7 +311,7 @@ macro_rules! possible_values_completer {
     };
 }
 
-possible_values_completer!(setoptions_values_completer, crate::cmd::SET_OPTS);
+possible_values_completer!(setoptions_values_completer, crate::command::click::SET_OPTS);
 
 possible_values_completer!(
     portforwardaction_values_completer,
